@@ -1,5 +1,4 @@
 package com.example.dummyjson;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -8,5 +7,6 @@ public class DummyJsonClientApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(DummyJsonClientApplication.class, args);
+        
     }
 }
